+++ tests/browser/karma.conf.esm.js
@@ -3,10 +3,10 @@ const karmaConfig = require('./karma-shared.conf.js');
 module.exports = function(config) {
   karmaConfig.logLevel = config.LOG_INFO;
 
-  karmaConfig.files.push('./tests/esm.spec.js');
+  karmaConfig.files.push('./modules/esm.js');
 
   karmaConfig.preprocessors = {
-    './tests/esm.spec.js': ['webpack']
+    './modules/esm.js': ['webpack']
   };
 
   config.set(karmaConfig);